continue factoring out scoring module
