+++ app/modules/scoring/effect-repository.ts
@@ -1,4 +1,5 @@
 import { type Transaction } from 'kysely'
+import { type DBEffect } from '#app/types/db-types.ts'
 import { type DB } from '#app/types/kysely-types.ts'
 import { snakeToCamelCaseObject } from './scoring-utils.ts'
 
@@ -16,4 +17,36 @@ export async function insertEffectEvent(trx: Transaction<DB>, data: any) {
 			oc.columns(['voteEventId', 'postId', 'commentId']).doNothing(),
 		)
 		.execute()
+}
+
+export async function getEffect(
+	trx: Transaction<DB>,
+	postId: number,
+	commentId: number,
+): Promise<DBEffect | undefined> {
+	const effect: DBEffect | undefined = await trx
+		.selectFrom('Effect')
+		.where('postId', '=', postId)
+		.where('commentId', '=', commentId)
+		.selectAll()
+		.executeTakeFirst()
+	return effect
+}
+
+export async function getEffects(
+	trx: Transaction<DB>,
+	postId: number,
+): Promise<DBEffect[]> {
+	let query = trx
+		.selectFrom('Post')
+		.innerJoin('EffectWithDefault as Effect', join =>
+			join.on('Effect.commentId', '=', postId),
+		)
+		.innerJoin('Post as TargetPost', 'TargetPost.id', 'Effect.postId')
+		.selectAll('Effect')
+		.where('Post.id', '=', postId)
+		.orderBy('TargetPost.createdAt')
+
+	const effects = await query.execute()
+	return effects
 }
