+++ app/repositories/ranking.ts
@@ -1,6 +1,7 @@
 import * as Immutable from 'immutable'
 import { type Transaction, sql } from 'kysely'
 import { MAX_POSTS_PER_PAGE } from '#app/constants.ts'
+import { getEffect } from '#app/modules/scoring/effect-repository.ts'
 import { getArtefact } from '#app/repositories/artefact.ts'
 import {
 	type Effect,
@@ -13,7 +14,6 @@ import {
 	type PollPagePost,
 } from '#app/types/api-types.ts'
 import { invariant } from '#app/utils/misc.tsx'
-import { type DBEffect } from '../types/db-types.ts'
 import { type DB } from '../types/kysely-types.ts'
 import { relativeEntropy } from '../utils/entropy.ts'
 import {
@@ -204,38 +204,6 @@ export function effectSizeOnTarget(effectOnTarget: Effect | null): number {
 	const targetQ = effectOnTarget?.q ?? 0
 	const targetPSize = effectOnTarget?.pSize ?? 0
 	return relativeEntropy(targetP, targetQ) * targetPSize
-}
-
-export async function getEffect(
-	trx: Transaction<DB>,
-	postId: number,
-	commentId: number,
-): Promise<DBEffect | undefined> {
-	const effect: DBEffect | undefined = await trx
-		.selectFrom('Effect')
-		.where('postId', '=', postId)
-		.where('commentId', '=', commentId)
-		.selectAll()
-		.executeTakeFirst()
-	return effect
-}
-
-export async function getEffects(
-	trx: Transaction<DB>,
-	postId: number,
-): Promise<DBEffect[]> {
-	let query = trx
-		.selectFrom('Post')
-		.innerJoin('EffectWithDefault as Effect', join =>
-			join.on('Effect.commentId', '=', postId),
-		)
-		.innerJoin('Post as TargetPost', 'TargetPost.id', 'Effect.postId')
-		.selectAll('Effect')
-		.where('Post.id', '=', postId)
-		.orderBy('TargetPost.createdAt')
-
-	const effects = await query.execute()
-	return effects
 }
 
 export async function getChronologicalToplevelPosts(
