+++ app/routes/stats.$postId.tsx
@@ -5,8 +5,8 @@ import { z } from 'zod'
 import { GeneralErrorBoundary } from '#app/components/error-boundary.tsx'
 import { Markdown } from '#app/components/markdown.tsx'
 import { db } from '#app/db.ts'
+import { getEffects } from '#app/modules/scoring/effect-repository.ts'
 import { getStatsPost } from '#app/repositories/post.ts'
-import { getEffects } from '#app/repositories/ranking.ts'
 import { type StatsPost } from '#app/types/api-types.ts'
 import { type DBEffect } from '#app/types/db-types.ts'
 import { relativeEntropy } from '#app/utils/entropy.ts'
